Add user role-based navigation and streamline UI
Implemented user role-based navigation to differentiate admin and normal user accounts. In the Login.js file, user data now includes an "isAdmin" flag which is stored in sessionStorage. If the user is an admin, navigation directs to the Account page; otherwise, it redirects to the Home page. Simplified the UI by removing unnecessary buttons from NavBarLinks.js and updated route in ActionButtons.js for user accounts.